+++ c-api/lib.rs
@@ -69,11 +69,8 @@ pub enum ttfp_glyph_class {
 
 /// @brief A glyph image format.
 #[repr(C)]
-pub enum ttfp_image_format {
+pub enum ttfp_raster_image_format {
     PNG = 0,
-    JPEG,
-    TIFF,
-    SVG,
 }
 
 /// @brief A glyph image.
@@ -81,7 +78,7 @@ pub enum ttfp_image_format {
 /// An image offset and size isn't defined in all tables, so `x`, `y`, `width` and `height`
 /// can be set to 0.
 #[repr(C)]
-pub struct ttfp_glyph_image {
+pub struct ttfp_glyph_raster_image {
     /// Horizontal offset.
     pub x: i16,
 
@@ -102,7 +99,7 @@ pub struct ttfp_glyph_image {
     pub pixels_per_em: u16,
 
     /// An image format.
-    pub format: ttfp_image_format,
+    pub format: ttfp_raster_image_format,
 
     /// A raw image data as is. It's up to the caller to decode PNG, JPEG, etc.
     pub data: *const c_char,
@@ -136,26 +133,39 @@ pub extern "C" fn ttfp_fonts_in_collection(data: *const c_char, len: usize) -> i
 
 /// @brief Creates a new font parser.
 ///
-/// This is the only heap allocation in the library.
+/// Since #ttfp_font is an opaque pointer, a caller should allocate it manually
+/// using #ttfp_font_size_of.
+/// Deallocation is also handled by a caller.
+/// #ttfp_font doesn't use heap internally, so we can simply `free()` it without
+/// a dedicated `ttfp_font_deinit` function.
 ///
-/// @param data The font data. Must outlive the #ttfp_font.
-/// @param len The size of the font data.
+/// @param data A font binary data. Must outlive the #ttfp_font.
+/// @param len Size of the font data.
 /// @param index The font index in a collection (typically *.ttc). 0 should be used for basic fonts.
-/// @return Font handle or NULL on error.
+/// @param font A pointer to a #ttfp_font object.
+/// @return `true` on success.
 #[no_mangle]
-pub extern "C" fn ttfp_create_font(data: *const c_char, len: usize, index: u32) -> *mut ttfp_font {
+pub extern "C" fn ttfp_font_init(data: *const c_char, len: usize, index: u32, font: *mut c_void) -> bool {
     // This method invokes a lot of parsing, so let's catch any panics just in case.
     std::panic::catch_unwind(|| {
         let data = unsafe { std::slice::from_raw_parts(data as *const _, len) };
-        let font = ttf_parser::Font::from_data(data, index).unwrap();
-        Box::into_raw(Box::new(font)) as *mut _
-    }).unwrap_or(std::ptr::null_mut())
+        let font_rs = ttf_parser::Font::from_data(data, index).unwrap();
+        unsafe {
+            std::ptr::copy(
+                &font_rs as *const ttf_parser::Font as _,
+                font,
+                ttfp_font_size_of(),
+            );
+        }
+
+        true
+    }).unwrap_or(false)
 }
 
-/// @brief Destroys the #ttfp_font.
+/// @brief Returns the size of `ttfp_font`.
 #[no_mangle]
-pub extern "C" fn ttfp_destroy_font(font: *mut ttfp_font) {
-    unsafe { Box::from_raw(font) };
+pub extern "C" fn ttfp_font_size_of() -> usize {
+    std::mem::size_of::<ttf_parser::Font>()
 }
 
 /// @brief Checks that font has a specified table.
@@ -674,51 +684,79 @@ pub extern "C" fn ttfp_get_glyph_bbox(
     }).unwrap_or(false)
 }
 
-/// @brief Returns a reference to a glyph image.
+/// @brief Returns a reference to a glyph's raster image.
 ///
 /// A font can define a glyph using a raster or a vector image instead of a simple outline.
-/// Which is primarily used for emojis. This method should be used to access those images.
+/// Which is primarily used for emojis. This method should be used to access raster images.
 ///
-/// `pixels_per_em` allows selecting a preferred image size. While the chosen size will
+/// `pixels_per_em` allows selecting a preferred image size. The chosen size will
 /// be closer to an upper one. So when font has 64px and 96px images and `pixels_per_em`
 /// is set to 72, 96px image will be returned.
 /// To get the largest image simply use `SHRT_MAX`.
-/// This property has no effect in case of SVG.
 ///
 /// Note that this method will return an encoded image. It should be decoded
-/// (in case of PNG, JPEG, etc.), rendered (in case of SVG) or even decompressed
-/// (in case of SVGZ) by the caller. We don't validate or preprocess it in any way.
+/// by the caller. We don't validate or preprocess it in any way.
 ///
-/// Also, a font can contain both: images and outlines. So when this method returns `None`
+/// Currently, only PNG images are supported.
+///
+/// Also, a font can contain both: images and outlines. So when this method returns `false`
 /// you should also try `ttfp_outline_glyph()` afterwards.
 ///
 /// There are multiple ways an image can be stored in a TrueType font
-/// and we support `sbix`, `CBLC`+`CBDT` and `SVG`.
+/// and this method supports only `sbix`, `CBLC`+`CBDT`.
 #[no_mangle]
-pub extern "C" fn ttfp_get_glyph_image(
+pub extern "C" fn ttfp_get_glyph_raster_image(
     font: *const ttfp_font,
     glyph_id: GlyphId,
     pixels_per_em: u16,
-    glyph_image: *mut ttfp_glyph_image,
+    glyph_image: *mut ttfp_glyph_raster_image,
 ) -> bool {
-    match font_from_ptr(font).glyph_image(glyph_id, pixels_per_em) {
+    match font_from_ptr(font).glyph_raster_image(glyph_id, pixels_per_em) {
         Some(image) => {
             unsafe {
-                *glyph_image = ttfp_glyph_image {
-                    x: image.x.unwrap_or(0),
-                    y: image.y.unwrap_or(0),
-                    width: image.width.unwrap_or(0),
-                    height: image.height.unwrap_or(0),
+                *glyph_image = ttfp_glyph_raster_image {
+                    x: image.x,
+                    y: image.y,
+                    width: image.width,
+                    height: image.height,
                     pixels_per_em: image.pixels_per_em,
                     format: match image.format {
-                        ttf_parser::ImageFormat::PNG => ttfp_image_format::PNG,
-                        ttf_parser::ImageFormat::JPEG => ttfp_image_format::JPEG,
-                        ttf_parser::ImageFormat::TIFF => ttfp_image_format::TIFF,
-                        ttf_parser::ImageFormat::SVG => ttfp_image_format::SVG,
+                        ttf_parser::RasterImageFormat::PNG => ttfp_raster_image_format::PNG,
                     },
                     data: image.data.as_ptr() as _,
                     len: image.data.len() as u32,
                 };
+            }
+
+            true
+        }
+        None => false,
+    }
+}
+
+/// @brief Returns a reference to a glyph's SVG image.
+///
+/// A font can define a glyph using a raster or a vector image instead of a simple outline.
+/// Which is primarily used for emojis. This method should be used to access SVG images.
+///
+/// Note that this method will return just an SVG data. It should be rendered
+/// or even decompressed (in case of SVGZ) by the caller.
+/// We don't validate or preprocess it in any way.
+///
+/// Also, a font can contain both: images and outlines. So when this method returns `false`
+/// you should also try `ttfp_outline_glyph()` afterwards.
+#[no_mangle]
+pub extern "C" fn ttfp_get_glyph_svg_image(
+    font: *const ttfp_font,
+    glyph_id: GlyphId,
+    svg: *mut *const c_char,
+    len: *mut u32,
+) -> bool {
+    match font_from_ptr(font).glyph_svg_image(glyph_id) {
+        Some(image) => {
+            unsafe {
+                *svg = image.as_ptr() as *const c_char;
+                *len = image.len() as u32;
             }
 
             true